pyproject.toml
[tool.poetry]
name = "finbot"
version = "0.1.0"
description = ""
authors = ["egor187 <py.egor.py@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.14.3"
python-dotenv = "^0.19.0"
pytz = "^2021.1"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
